Remove triton dependency from FMS
